requirements.txt
flask
langchain
langchain_community
langchain_text-splitters
langchain_openai
langchain_core
sentence_transformers
chromadb
python_dotenv
unstructured[all-docs]
poppler-utils
werkzeug
markdown